separate Search Devices UAT tests to different file
